Separate SNAP variables into their own files
