+++ openfisca_us/variables/usda/snap/income/deductions/snap_standard_deduction.py
